+++ internal/core/domain/model.go
@@ -2,6 +2,14 @@ package domain
 
 import (
 	"github.com/google/uuid"
+	"os"
+	"os/user"
+	"path/filepath"
+)
+
+var (
+	_defaultOutputPath        = "Videos"
+	_defaultRecordsFolderPath = "Best Records"
 )
 
 type (
@@ -21,3 +29,19 @@ type (
 		Streams []Stream
 	}
 )
+
+func GetOutputPath() (string, error) {
+	currentUser, err := user.Current()
+	if err != nil {
+		return "", err
+	}
+
+	outputPath := filepath.Join(currentUser.HomeDir, _defaultOutputPath, _defaultRecordsFolderPath)
+	if _, err := os.Stat(outputPath); os.IsNotExist(err) {
+		err := os.MkdirAll(outputPath, 0755) // Permissões rwxr-xr-x
+		if err != nil {
+			return "", err
+		}
+	}
+	return outputPath, nil
+}